reposition RevenueViewModel to ViewModels
